% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapplyFunGen.R
\docType{package}
\name{mclapplyFunGen-package}
\title{mclapplyFunGen}
\description{
Generates an mclapply function.
}
\author{
\strong{Maintainer}: Dale \email{dale@piratepress.org} [copyright holder]

}
