% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapplyFunGen.R
\name{mclapplyFunGen}
\alias{mclapplyFunGen}
\title{mclapplyFunGen}
\usage{
mclapplyFunGen(
  mc.preschedule = TRUE,
  mc.set.seed = FALSE,
  mc.silent = TRUE,
  mc.cores = getLapply::getSensibleThreads()
)
}
\description{
mclapplyFunGen
}
